separating classes / reworked move
