+++ hpTypes.js
@@ -1,0 +1,22 @@
+// Hidden Power Types for Type calculation on Pokemon Generation
+
+const hpTypes = [
+  "Fighting",
+  "Flying",
+  "Poison",
+  "Ground",
+  "Rock",
+  "Bug",
+  "Ghost",
+  "Steel",
+  "Fire",
+  "Water",
+  "Grass",
+  "Electric",
+  "Psychic",
+  "Ice",
+  "Dragon",
+  "Dark"
+]
+
+module.exports.hpTypes = hpTypes;
